#19 Refactor: 인터페이스와 구현체를 분리한다
